Refactor string utilities into strings package
